Standardized all thresholds to 0-255, no more floating point thresholds
Ditz-issue: 961ee390856d097742ea1dab3cb79ca6c64288e1

+++ include/code_grid.h
@@ -38,7 +38,7 @@ typedef struct {
 
 
 
-void koki_grid_from_IplImage(IplImage *unwarped_frame, float threshold,
+void koki_grid_from_IplImage(IplImage *unwarped_frame, uint16_t threshold,
 			     koki_grid_t *grid);
 
 void koki_grid_print(koki_grid_t *grid);
